Fix #EFORGE-5 and introduce new action tracking sytem. This is a schema change.
The new action tracking is significantly more powerful - For example, it
tracks both the old and new values of a field, and the renderers do
clever things too like displaying what was removed and added to the
dependencies field.

We have imported django-picklefiled into eforge.util.picklefield in order
to store this. Yes, this does unfortunately mean our actions are a difficult
to understand blob, but the alternatives are worse. Fortuately, the pickled
data should be pretty simple: Lists, integers and tuples.